{{cookiecutter.github_repository}}/{{cookiecutter.backend_path}}/requirements/production.txt
# Pro-tip: Try not to put anything here. There should be no dependency in
#   production that isn't in development.
-r common.txt

# Static Files and Media Storage
# -------------------------------------
google-cloud-storage==1.20.0
django-storages[google]==1.7.2

# Caching
# -------------------------------------
django-redis==4.10.0
hiredis==1.0.0


{%- if cookiecutter.add_django_auth_wall.lower() == 'y' %}
# Basic Auth Protection
# -------------------------------------
django-auth-wall==0.2.0
{%- endif %}

# server
uwsgi==2.0.18
